Feature/73 default json mapper response (#201)
* feat: added customizable default value for jsonMapper

* reafactor: better typings and format
