+++ packages/xlsx-import/tests/unit/mappers/stringMapper.test.ts
@@ -1,5 +1,5 @@
 import * as chai from 'chai';
-import { isEmpty, isFilled, jsonMapper, lowerCaseMapper, upperCaseMapper } from '../../../src/mappers';
+import { isEmpty, isFilled, lowerCaseMapper, upperCaseMapper } from '../../../src/mappers';
 import { stringMapper } from '../../../src/mappers/stringMapper';
 
 describe('UNIT TEST: src/mappers/', () => {
@@ -58,30 +58,6 @@ describe('UNIT TEST: src/mappers/', () => {
         dataProvider.forEach(({ inValue, expectedResult }) => {
             it(`lowerCaseMapper for input "${inValue}" SHOULD return "${expectedResult}"`, () => {
                 chai.expect(lowerCaseMapper(inValue)).equal(expectedResult);
-            });
-        });
-    });
-
-    // todo move to jsonMapper.test.ts
-    describe('jsonMapper', () => {
-        const dataProvider = [
-            { inValue: '', expectedResult: null },
-            { inValue: null, expectedResult: null },
-            { inValue: ' ', expectedResult: null },
-            { inValue: 'asd', expectedResult: null },
-            { inValue: 'null', expectedResult: null },
-            { inValue: 'false', expectedResult: false },
-            { inValue: 'true', expectedResult: true },
-            { inValue: '0', expectedResult: 0 },
-            { inValue: '1', expectedResult: 1 },
-            { inValue: '"string"', expectedResult: 'string' },
-            { inValue: "'string'", expectedResult: null },
-            { inValue: '{"a":1}', expectedResult: { a: 1 } },
-            { inValue: '{"a":[1]}', expectedResult: { a: [1] } },
-        ];
-        dataProvider.forEach(({ inValue, expectedResult }) => {
-            it(`jsonMapper for input "${inValue}" SHOULD return "${JSON.stringify(expectedResult)}"`, () => {
-                chai.expect(jsonMapper(inValue as string)).is.eql(expectedResult);
             });
         });
     });
